+++ src/webclient/GameStateProvider.ts
@@ -90,7 +90,7 @@ export class GameStateProvider {
     this.channel.push("avatar_update", message);
   }
 
-  public sendPlayerAction(actionType: string, targetId: string): void {
-    this.channel.push("player_action", { action: actionType, target_id: targetId });
+  public sendPlayerAction(actionType: string, targetId: string, metadata?: any): void {
+    this.channel.push("player_action", { action: actionType, target_id: targetId, metadata });
   }
 }